+++ src/flow/src/batching_mode/task.rs
@@ -28,10 +28,8 @@ use datafusion::datasource::DefaultTableSource;
 use datafusion::sql::unparser::expr_to_sql;
 use datafusion_common::DFSchemaRef;
 use datafusion_common::tree_node::{Transformed, TreeNode};
-use datafusion_expr::{DmlStatement, Expr, LogicalPlan, WriteOp};
-use datatypes::prelude::ConcreteDataType;
-use datatypes::schema::{ColumnSchema, Schema};
-use operator::expr_helper::column_schemas_to_defs;
+use datafusion_expr::{DmlStatement, LogicalPlan, WriteOp};
+use datatypes::schema::Schema;
 use query::QueryEngineRef;
 use query::options::{
     FLOW_INCREMENTAL_AFTER_SEQS, FLOW_INCREMENTAL_MODE, FLOW_INCREMENTAL_MODE_MEMTABLE_ONLY,
@@ -48,20 +46,24 @@ use tokio::sync::oneshot;
 use tokio::sync::oneshot::error::TryRecvError;
 use tokio::time::Instant;
 
-use crate::adapter::{AUTO_CREATED_PLACEHOLDER_TS_COL, AUTO_CREATED_UPDATE_AT_TS_COL};
 use crate::batching_mode::BatchingModeOptions;
+use crate::batching_mode::checkpoint::{
+    FlowCheckpointDecision, FlowQueryFallbackReason, checkpoint_mode_label,
+};
 use crate::batching_mode::frontend_client::FrontendClient;
+use crate::batching_mode::incremental_filter::build_sink_dirty_time_window_filter_expr;
 use crate::batching_mode::state::{CheckpointMode, FilterExprInfo, TaskState};
+use crate::batching_mode::table_creator::{QueryType, create_table_with_expr};
 use crate::batching_mode::time_window::TimeWindowExpr;
 use crate::batching_mode::utils::{
-    AddFilterRewriter, ColumnMatcherRewriter, FindGroupByFinalName, IncrementalAggregateAnalysis,
-    analyze_incremental_aggregate_plan, gen_plan_with_matching_schema, get_table_info_df_schema,
+    AddFilterRewriter, ColumnMatcherRewriter, analyze_incremental_aggregate_plan,
+    gen_plan_with_matching_schema, get_table_info_df_schema,
     rewrite_incremental_aggregate_with_sink_merge, sql_to_df_plan,
 };
 use crate::df_optimizer::apply_df_optimizer;
 use crate::error::{
-    ConvertColumnSchemaSnafu, DatafusionSnafu, ExternalSnafu, InvalidQuerySnafu,
-    SubstraitEncodeLogicalPlanSnafu, UnexpectedSnafu,
+    DatafusionSnafu, ExternalSnafu, InvalidQuerySnafu, SubstraitEncodeLogicalPlanSnafu,
+    UnexpectedSnafu,
 };
 use crate::metrics::{
     METRIC_FLOW_BATCHING_ENGINE_CHECKPOINT_DECISION_CNT, METRIC_FLOW_BATCHING_ENGINE_ERROR_CNT,
@@ -70,10 +72,6 @@ use crate::metrics::{
     METRIC_FLOW_ROWS,
 };
 use crate::{Error, FlowId};
-
-const CHECKPOINT_DECISION_ADVANCE: &str = "advance";
-const CHECKPOINT_DECISION_FALLBACK: &str = "fallback";
-const CHECKPOINT_REASON_NONE: &str = "none";
 
 /// The task's config, immutable once created
 #[derive(Clone)]
@@ -93,108 +91,6 @@ pub struct TaskConfig {
     pub flow_eval_interval: Option<Duration>,
 }
 
-/// Why the task fell back to full snapshot mode.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum FlowQueryFallbackReason {
-    /// The query result did not include a region-watermark map at all.
-    MissingRegionWatermark,
-    /// Some participating regions could not prove safe advancement against
-    /// both the returned watermarks and the checkpoint map.
-    IncompleteRegionWatermark,
-    /// The datanode detected a stale incremental cursor and the Flow
-    /// must recompute from scratch.
-    StaleCursor,
-    /// A non-stale-cursor query failure; the Flow resets to full snapshot
-    /// to avoid cascading errors.
-    IncrementalQueryFailure,
-    /// Incremental mode has been permanently disabled for this Flow
-    /// (e.g. because the query shape is not incrementally safe).
-    IncrementalDisabled,
-}
-
-impl FlowQueryFallbackReason {
-    fn as_label(self) -> &'static str {
-        match self {
-            Self::MissingRegionWatermark => "missing_region_watermark",
-            Self::IncompleteRegionWatermark => "incomplete_region_watermark",
-            Self::StaleCursor => "stale_cursor",
-            Self::IncrementalQueryFailure => "incremental_query_failure",
-            Self::IncrementalDisabled => "incremental_disabled",
-        }
-    }
-}
-
-/// Decision produced by [`BatchingTask::apply_query_result_to_state`] after
-/// each Flow query execution. Describes whether the task advanced its
-/// checkpoint state or fell back to full snapshot, and why.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum FlowCheckpointDecision {
-    /// FullSnapshot → Incremental transition.
-    ///
-    /// The query exercised every participating region, all returned valid
-    /// watermarks, and the checkpoint map was populated from scratch.
-    /// Subsequent executions will use incremental after-seqs.
-    AdvancedFromFullSnapshot {
-        participating_regions: usize,
-        watermarks: usize,
-    },
-    /// Existing Incremental → Incremental (in-place advancement).
-    ///
-    /// A subset of participating regions advanced their watermarks. The
-    /// task stays in incremental mode with an updated checkpoint map.
-    AdvancedIncremental {
-        participating_regions: usize,
-        watermarks: usize,
-    },
-    /// Any mode → FullSnapshot.
-    ///
-    /// Watermark information was incomplete, a participating region was
-    /// absent from the existing checkpoint map, the task has permanently
-    /// disabled incremental mode, or the query itself failed. The task
-    /// resets to full snapshot semantics for the next execution.
-    FallbackToFullSnapshot {
-        previous_mode: CheckpointMode,
-        reason: FlowQueryFallbackReason,
-    },
-}
-
-impl FlowCheckpointDecision {
-    fn mode_label(self) -> &'static str {
-        match self {
-            Self::AdvancedFromFullSnapshot { .. } => {
-                checkpoint_mode_label(CheckpointMode::FullSnapshot)
-            }
-            Self::AdvancedIncremental { .. } => checkpoint_mode_label(CheckpointMode::Incremental),
-            Self::FallbackToFullSnapshot { previous_mode, .. } => {
-                checkpoint_mode_label(previous_mode)
-            }
-        }
-    }
-
-    fn decision_label(self) -> &'static str {
-        match self {
-            Self::AdvancedFromFullSnapshot { .. } | Self::AdvancedIncremental { .. } => {
-                CHECKPOINT_DECISION_ADVANCE
-            }
-            Self::FallbackToFullSnapshot { .. } => CHECKPOINT_DECISION_FALLBACK,
-        }
-    }
-
-    fn reason_label(self) -> &'static str {
-        match self {
-            Self::FallbackToFullSnapshot { reason, .. } => reason.as_label(),
-            _ => CHECKPOINT_REASON_NONE,
-        }
-    }
-}
-
-fn checkpoint_mode_label(mode: CheckpointMode) -> &'static str {
-    match mode {
-        CheckpointMode::FullSnapshot => "full_snapshot",
-        CheckpointMode::Incremental => "incremental",
-    }
-}
-
 fn determine_query_type(query: &str, query_ctx: &QueryContextRef) -> Result<QueryType, Error> {
     let is_tql = is_tql(query_ctx.sql_dialect(), query)
         .map_err(BoxedError::new)
@@ -211,14 +107,6 @@ fn is_merge_mode_last_non_null(options: &HashMap<String, String>) -> bool {
         .get(MERGE_MODE_KEY)
         .map(|mode| mode.eq_ignore_ascii_case("last_non_null"))
         .unwrap_or(false)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum QueryType {
-    /// query is a tql query
-    Tql,
-    /// query is a sql query
-    Sql,
 }
 
 #[derive(Clone)]
@@ -1309,1087 +1197,5 @@ impl BatchingTask {
     }
 }
 
-fn build_sink_dirty_time_window_filter_expr(
-    flow_id: FlowId,
-    analysis: &IncrementalAggregateAnalysis,
-    sink_schema: &Schema,
-    dirty_filter: Option<&FilterExprInfo>,
-) -> Result<Option<Expr>, Error> {
-    let Some(dirty_filter) = dirty_filter else {
-        return Ok(None);
-    };
-
-    let Some(sink_filter_col) =
-        infer_sink_time_window_filter_col(flow_id, analysis, sink_schema, dirty_filter)
-    else {
-        return Ok(None);
-    };
-
-    dirty_filter.predicate_for_col(&sink_filter_col)
-}
-
-fn infer_sink_time_window_filter_col(
-    flow_id: FlowId,
-    analysis: &IncrementalAggregateAnalysis,
-    sink_schema: &Schema,
-    dirty_filter: &FilterExprInfo,
-) -> Option<String> {
-    if analysis.group_key_names.is_empty() {
-        return None;
-    }
-
-    let is_timestamp_group_key = |name: &str| {
-        analysis.group_key_names.iter().any(|key| key == name)
-            && sink_schema
-                .column_schema_by_name(name)
-                .is_some_and(|col| col.data_type.is_timestamp())
-    };
-
-    if is_timestamp_group_key(&dirty_filter.col_name) {
-        return Some(dirty_filter.col_name.clone());
-    }
-
-    let candidates = analysis
-        .group_key_names
-        .iter()
-        .filter(|name| is_timestamp_group_key(name))
-        .cloned()
-        .collect::<Vec<_>>();
-
-    match candidates.as_slice() {
-        [name] => Some(name.clone()),
-        [] => {
-            warn!(
-                "Flow {} cannot infer sink dirty-window filter column: no timestamp group key in {:?}",
-                flow_id, analysis.group_key_names
-            );
-            None
-        }
-        _ => {
-            warn!(
-                "Flow {} cannot infer sink dirty-window filter column: ambiguous timestamp group keys {:?}",
-                flow_id, candidates
-            );
-            None
-        }
-    }
-}
-
-// auto created table have a auto added column `update_at`, and optional have a `AUTO_CREATED_PLACEHOLDER_TS_COL` column for time index placeholder if no timestamp column is specified
-// TODO(discord9): for now no default value is set for auto added column for compatibility reason with streaming mode, but this might change in favor of simpler code?
-fn create_table_with_expr(
-    plan: &LogicalPlan,
-    sink_table_name: &[String; 3],
-    query_type: &QueryType,
-) -> Result<CreateTableExpr, Error> {
-    let table_def = match query_type {
-        &QueryType::Sql => {
-            if let Some(def) = build_pk_from_aggr(plan)? {
-                def
-            } else {
-                build_by_sql_schema(plan)?
-            }
-        }
-        QueryType::Tql => {
-            // first try build from aggr, then from tql schema because tql query might not have aggr node
-            if let Some(table_def) = build_pk_from_aggr(plan)? {
-                table_def
-            } else {
-                build_by_tql_schema(plan)?
-            }
-        }
-    };
-    let first_time_stamp = table_def.ts_col;
-    let primary_keys = table_def.pks;
-
-    let mut column_schemas = Vec::new();
-    for field in plan.schema().fields() {
-        let name = field.name();
-        let ty = ConcreteDataType::from_arrow_type(field.data_type());
-        let col_schema = if first_time_stamp == Some(name.clone()) {
-            ColumnSchema::new(name, ty, false).with_time_index(true)
-        } else {
-            ColumnSchema::new(name, ty, true)
-        };
-
-        match query_type {
-            QueryType::Sql => {
-                column_schemas.push(col_schema);
-            }
-            QueryType::Tql => {
-                // if is val column, need to rename as val DOUBLE NULL
-                // if is tag column, need to cast type as STRING NULL
-                let is_tag_column = primary_keys.contains(name);
-                let is_val_column = !is_tag_column && first_time_stamp.as_ref() != Some(name);
-                if is_val_column {
-                    let col_schema =
-                        ColumnSchema::new(name, ConcreteDataType::float64_datatype(), true);
-                    column_schemas.push(col_schema);
-                } else if is_tag_column {
-                    let col_schema =
-                        ColumnSchema::new(name, ConcreteDataType::string_datatype(), true);
-                    column_schemas.push(col_schema);
-                } else {
-                    // time index column
-                    column_schemas.push(col_schema);
-                }
-            }
-        }
-    }
-
-    if query_type == &QueryType::Sql {
-        let update_at_schema = ColumnSchema::new(
-            AUTO_CREATED_UPDATE_AT_TS_COL,
-            ConcreteDataType::timestamp_millisecond_datatype(),
-            true,
-        );
-        column_schemas.push(update_at_schema);
-    }
-
-    let time_index = if let Some(time_index) = first_time_stamp {
-        time_index
-    } else {
-        column_schemas.push(
-            ColumnSchema::new(
-                AUTO_CREATED_PLACEHOLDER_TS_COL,
-                ConcreteDataType::timestamp_millisecond_datatype(),
-                false,
-            )
-            .with_time_index(true),
-        );
-        AUTO_CREATED_PLACEHOLDER_TS_COL.to_string()
-    };
-
-    let column_defs =
-        column_schemas_to_defs(column_schemas, &primary_keys).context(ConvertColumnSchemaSnafu)?;
-    Ok(CreateTableExpr {
-        catalog_name: sink_table_name[0].clone(),
-        schema_name: sink_table_name[1].clone(),
-        table_name: sink_table_name[2].clone(),
-        desc: "Auto created table by flow engine".to_string(),
-        column_defs,
-        time_index,
-        primary_keys,
-        create_if_not_exists: true,
-        table_options: Default::default(),
-        table_id: None,
-        engine: "mito".to_string(),
-    })
-}
-
-/// simply build by schema, return first timestamp column and no primary key
-fn build_by_sql_schema(plan: &LogicalPlan) -> Result<TableDef, Error> {
-    let first_time_stamp = plan.schema().fields().iter().find_map(|f| {
-        if ConcreteDataType::from_arrow_type(f.data_type()).is_timestamp() {
-            Some(f.name().clone())
-        } else {
-            None
-        }
-    });
-    Ok(TableDef {
-        ts_col: first_time_stamp,
-        pks: vec![],
-    })
-}
-
-/// Return first timestamp column found in output schema and all string columns
-fn build_by_tql_schema(plan: &LogicalPlan) -> Result<TableDef, Error> {
-    let first_time_stamp = plan.schema().fields().iter().find_map(|f| {
-        if ConcreteDataType::from_arrow_type(f.data_type()).is_timestamp() {
-            Some(f.name().clone())
-        } else {
-            None
-        }
-    });
-    let string_columns = plan
-        .schema()
-        .fields()
-        .iter()
-        .filter_map(|f| {
-            if ConcreteDataType::from_arrow_type(f.data_type()).is_string() {
-                Some(f.name().clone())
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-
-    Ok(TableDef {
-        ts_col: first_time_stamp,
-        pks: string_columns,
-    })
-}
-
-struct TableDef {
-    ts_col: Option<String>,
-    pks: Vec<String>,
-}
-
-/// Return first timestamp column which is in group by clause and other columns which are also in group by clause
-///
-/// # Returns
-///
-/// * `Option<String>` - first timestamp column which is in group by clause
-/// * `Vec<String>` - other columns which are also in group by clause
-///
-/// if no aggregation found, return None
-fn build_pk_from_aggr(plan: &LogicalPlan) -> Result<Option<TableDef>, Error> {
-    let fields = plan.schema().fields();
-    let mut pk_names = FindGroupByFinalName::default();
-
-    plan.visit(&mut pk_names)
-        .with_context(|_| DatafusionSnafu {
-            context: format!("Can't find aggr expr in plan {plan:?}"),
-        })?;
-
-    // if no group by clause, return empty with first timestamp column found in output schema
-    let Some(pk_final_names) = pk_names.get_group_expr_names() else {
-        return Ok(None);
-    };
-    if pk_final_names.is_empty() {
-        let first_ts_col = fields
-            .iter()
-            .find(|f| ConcreteDataType::from_arrow_type(f.data_type()).is_timestamp())
-            .map(|f| f.name().clone());
-        return Ok(Some(TableDef {
-            ts_col: first_ts_col,
-            pks: vec![],
-        }));
-    }
-
-    let all_pk_cols: Vec<_> = fields
-        .iter()
-        .filter(|f| pk_final_names.contains(f.name()))
-        .map(|f| f.name().clone())
-        .collect();
-    // auto create table use first timestamp column in group by clause as time index
-    let first_time_stamp = fields
-        .iter()
-        .find(|f| {
-            all_pk_cols.contains(&f.name().clone())
-                && ConcreteDataType::from_arrow_type(f.data_type()).is_timestamp()
-        })
-        .map(|f| f.name().clone());
-
-    let all_pk_cols: Vec<_> = all_pk_cols
-        .into_iter()
-        .filter(|col| first_time_stamp.as_ref() != Some(col))
-        .collect();
-
-    Ok(Some(TableDef {
-        ts_col: first_time_stamp,
-        pks: all_pk_cols,
-    }))
-}
-
 #[cfg(test)]
-mod test {
-    use std::collections::BTreeMap;
-
-    use api::v1::column_def::try_as_column_schema;
-    use common_error::ext::PlainError;
-    use common_error::status_code::StatusCode;
-    use common_query::Output;
-    use common_recordbatch::adapter::{RecordBatchMetrics, RegionWatermarkEntry};
-    use pretty_assertions::assert_eq;
-    use session::context::QueryContext;
-    use snafu::GenerateImplicitData;
-    use store_api::storage::consts::{
-        FLOW_STALE_CURSOR_ERROR_MARKER, FLOW_STALE_CURSOR_RETRY_HINT_FULL_RECOMPUTE,
-    };
-
-    use super::*;
-    use crate::test_utils::create_test_query_engine;
-
-    fn output_with_region_watermarks(
-        watermarks: impl IntoIterator<Item = (u64, Option<u64>)>,
-    ) -> OutputWithMetrics {
-        let result = OutputWithMetrics::from_output(Output::new_with_affected_rows(0));
-        result.metrics.update(Some(RecordBatchMetrics {
-            region_watermarks: watermarks
-                .into_iter()
-                .map(|(region_id, watermark)| RegionWatermarkEntry {
-                    region_id,
-                    watermark,
-                })
-                .collect(),
-            ..Default::default()
-        }));
-        result.metrics.mark_ready();
-        result
-    }
-
-    fn test_analysis_with_group_keys(group_key_names: Vec<&str>) -> IncrementalAggregateAnalysis {
-        IncrementalAggregateAnalysis {
-            group_key_names: group_key_names
-                .into_iter()
-                .map(|name| name.to_string())
-                .collect(),
-            merge_columns: vec![],
-            literal_columns: vec![],
-            output_field_names: vec![],
-            unsupported_exprs: vec![],
-        }
-    }
-
-    fn test_dirty_filter(col_name: &str) -> FilterExprInfo {
-        FilterExprInfo {
-            expr: datafusion_expr::col(col_name),
-            col_name: col_name.to_string(),
-            time_ranges: vec![],
-            window_size: chrono::Duration::seconds(1),
-        }
-    }
-
-    fn test_sink_schema(columns: Vec<(&str, ConcreteDataType)>) -> Schema {
-        Schema::new(
-            columns
-                .into_iter()
-                .map(|(name, data_type)| ColumnSchema::new(name, data_type, true))
-                .collect(),
-        )
-    }
-
-    #[test]
-    fn test_infer_sink_time_window_filter_col_uses_matching_source_group_key() {
-        let analysis = test_analysis_with_group_keys(vec!["ts", "host"]);
-        let sink_schema = test_sink_schema(vec![
-            ("ts", ConcreteDataType::timestamp_millisecond_datatype()),
-            ("host", ConcreteDataType::string_datatype()),
-        ]);
-        let dirty_filter = test_dirty_filter("ts");
-
-        assert_eq!(
-            Some("ts".to_string()),
-            infer_sink_time_window_filter_col(1, &analysis, &sink_schema, &dirty_filter)
-        );
-    }
-
-    #[test]
-    fn test_infer_sink_time_window_filter_col_uses_unique_timestamp_group_key() {
-        let analysis = test_analysis_with_group_keys(vec!["host", "time_window"]);
-        let sink_schema = test_sink_schema(vec![
-            ("host", ConcreteDataType::string_datatype()),
-            (
-                "time_window",
-                ConcreteDataType::timestamp_millisecond_datatype(),
-            ),
-            (
-                AUTO_CREATED_UPDATE_AT_TS_COL,
-                ConcreteDataType::timestamp_millisecond_datatype(),
-            ),
-        ]);
-        let dirty_filter = test_dirty_filter("ts");
-
-        assert_eq!(
-            Some("time_window".to_string()),
-            infer_sink_time_window_filter_col(1, &analysis, &sink_schema, &dirty_filter)
-        );
-    }
-
-    #[test]
-    fn test_infer_sink_time_window_filter_col_skips_global_aggregate() {
-        let analysis = test_analysis_with_group_keys(vec![]);
-        let sink_schema = test_sink_schema(vec![
-            ("number", ConcreteDataType::uint32_datatype()),
-            (
-                "time_window",
-                ConcreteDataType::timestamp_millisecond_datatype(),
-            ),
-        ]);
-        let dirty_filter = test_dirty_filter("ts");
-
-        assert_eq!(
-            None,
-            infer_sink_time_window_filter_col(1, &analysis, &sink_schema, &dirty_filter)
-        );
-    }
-
-    #[test]
-    fn test_infer_sink_time_window_filter_col_skips_without_timestamp_group_key() {
-        let analysis = test_analysis_with_group_keys(vec!["host", "device"]);
-        let sink_schema = test_sink_schema(vec![
-            ("host", ConcreteDataType::string_datatype()),
-            ("device", ConcreteDataType::string_datatype()),
-            (
-                AUTO_CREATED_UPDATE_AT_TS_COL,
-                ConcreteDataType::timestamp_millisecond_datatype(),
-            ),
-        ]);
-        let dirty_filter = test_dirty_filter("ts");
-
-        assert_eq!(
-            None,
-            infer_sink_time_window_filter_col(1, &analysis, &sink_schema, &dirty_filter)
-        );
-    }
-
-    #[test]
-    fn test_infer_sink_time_window_filter_col_skips_ambiguous_timestamp_group_keys() {
-        let analysis = test_analysis_with_group_keys(vec!["ts", "time_window"]);
-        let sink_schema = test_sink_schema(vec![
-            ("ts", ConcreteDataType::timestamp_millisecond_datatype()),
-            (
-                "time_window",
-                ConcreteDataType::timestamp_millisecond_datatype(),
-            ),
-        ]);
-        let dirty_filter = test_dirty_filter("source_ts");
-
-        assert_eq!(
-            None,
-            infer_sink_time_window_filter_col(1, &analysis, &sink_schema, &dirty_filter)
-        );
-    }
-
-    async fn new_test_task_with_missing_sink() -> BatchingTask {
-        new_test_task_and_plan_with_missing_sink().await.0
-    }
-
-    async fn new_test_task_and_plan_with_missing_sink() -> (BatchingTask, LogicalPlan) {
-        let query_engine = create_test_query_engine();
-        let ctx = QueryContext::arc();
-        let plan = sql_to_df_plan(
-            ctx.clone(),
-            query_engine.clone(),
-            "SELECT number, ts FROM numbers_with_ts",
-            true,
-        )
-        .await
-        .unwrap();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-
-        let task = BatchingTask::try_new(TaskArgs {
-            flow_id: 1,
-            query: "SELECT number, ts FROM numbers_with_ts",
-            plan: plan.clone(),
-            time_window_expr: None,
-            expire_after: None,
-            sink_table_name: [
-                "greptime".to_string(),
-                "public".to_string(),
-                "missing_sink".to_string(),
-            ],
-            source_table_names: vec![[
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ]],
-            query_ctx: ctx,
-            catalog_manager: query_engine.engine_state().catalog_manager().clone(),
-            shutdown_rx: rx,
-            batch_opts: Arc::new(BatchingModeOptions::default()),
-            flow_eval_interval: None,
-        })
-        .unwrap();
-
-        (task, plan)
-    }
-
-    fn ordinary_query_error() -> Error {
-        Error::External {
-            source: BoxedError::new(PlainError::new(
-                "ordinary query failure".to_string(),
-                StatusCode::EngineExecuteQuery,
-            )),
-            location: snafu::Location::generate(),
-        }
-    }
-
-    fn stale_cursor_query_error() -> Error {
-        Error::External {
-            source: BoxedError::new(PlainError::new(
-                format!(
-                    "{} region_id=1 given_seq=9 min_readable_seq=18 retry_hint={}",
-                    FLOW_STALE_CURSOR_ERROR_MARKER, FLOW_STALE_CURSOR_RETRY_HINT_FULL_RECOMPUTE
-                ),
-                StatusCode::EngineExecuteQuery,
-            )),
-            location: snafu::Location::generate(),
-        }
-    }
-
-    #[test]
-    fn test_apply_query_result_to_state_advances_full_snapshot_to_incremental() {
-        let query_ctx = QueryContext::arc();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let mut state = TaskState::new(query_ctx, rx);
-        let result = output_with_region_watermarks([(1_u64, Some(10_u64)), (2_u64, Some(20_u64))]);
-
-        let decision = BatchingTask::apply_query_result_to_state(
-            &mut state,
-            &result,
-            std::time::Duration::from_millis(1),
-        );
-
-        assert_eq!(
-            decision,
-            FlowCheckpointDecision::AdvancedFromFullSnapshot {
-                participating_regions: 2,
-                watermarks: 2,
-            }
-        );
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::Incremental);
-        assert_eq!(
-            state.checkpoints(),
-            &BTreeMap::from([(1_u64, 10_u64), (2_u64, 20_u64)])
-        );
-    }
-
-    #[test]
-    fn test_apply_query_result_to_state_stays_full_snapshot_when_incremental_disabled() {
-        let query_ctx = QueryContext::arc();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let mut state = TaskState::new(query_ctx, rx);
-        state.disable_incremental();
-        assert!(state.is_incremental_disabled());
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-
-        let result = output_with_region_watermarks([(1_u64, Some(10_u64)), (2_u64, Some(20_u64))]);
-        let decision = BatchingTask::apply_query_result_to_state(
-            &mut state,
-            &result,
-            std::time::Duration::from_millis(1),
-        );
-
-        // Should NOT claim advancement to incremental; should fallback with correct reason.
-        assert_eq!(
-            decision,
-            FlowCheckpointDecision::FallbackToFullSnapshot {
-                previous_mode: CheckpointMode::FullSnapshot,
-                reason: FlowQueryFallbackReason::IncrementalDisabled,
-            }
-        );
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert!(state.is_incremental_disabled());
-        // Checkpoints are still updated even if mode doesn't advance.
-        assert_eq!(
-            state.checkpoints(),
-            &BTreeMap::from([(1_u64, 10_u64), (2_u64, 20_u64)])
-        );
-    }
-
-    #[test]
-    fn test_apply_query_result_to_state_rejects_unproved_watermark() {
-        let query_ctx = QueryContext::arc();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let mut state = TaskState::new(query_ctx, rx);
-        let result = output_with_region_watermarks([(1_u64, Some(10_u64)), (2_u64, None)]);
-
-        let decision = BatchingTask::apply_query_result_to_state(
-            &mut state,
-            &result,
-            std::time::Duration::from_millis(1),
-        );
-
-        assert_eq!(
-            decision,
-            FlowCheckpointDecision::FallbackToFullSnapshot {
-                previous_mode: CheckpointMode::FullSnapshot,
-                reason: FlowQueryFallbackReason::IncompleteRegionWatermark,
-            }
-        );
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert!(state.checkpoints().is_empty());
-    }
-
-    #[test]
-    fn test_apply_query_result_to_state_reports_missing_watermark() {
-        let query_ctx = QueryContext::arc();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let mut state = TaskState::new(query_ctx, rx);
-        let result = OutputWithMetrics::from_output(Output::new_with_affected_rows(0));
-
-        let decision = BatchingTask::apply_query_result_to_state(
-            &mut state,
-            &result,
-            std::time::Duration::from_millis(1),
-        );
-
-        assert_eq!(
-            decision,
-            FlowCheckpointDecision::FallbackToFullSnapshot {
-                previous_mode: CheckpointMode::FullSnapshot,
-                reason: FlowQueryFallbackReason::MissingRegionWatermark,
-            }
-        );
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert!(state.checkpoints().is_empty());
-    }
-
-    #[test]
-    fn test_apply_query_result_to_state_advances_incremental_subset() {
-        let query_ctx = QueryContext::arc();
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let mut state = TaskState::new(query_ctx, rx);
-        state.advance_checkpoints(HashMap::from([
-            (1_u64, 10_u64),
-            (2_u64, 20_u64),
-            (3_u64, 30_u64),
-        ]));
-        let result = output_with_region_watermarks([(1_u64, Some(12_u64)), (3_u64, Some(35_u64))]);
-
-        let decision = BatchingTask::apply_query_result_to_state(
-            &mut state,
-            &result,
-            std::time::Duration::from_millis(1),
-        );
-
-        assert_eq!(
-            decision,
-            FlowCheckpointDecision::AdvancedIncremental {
-                participating_regions: 2,
-                watermarks: 2,
-            }
-        );
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::Incremental);
-        assert_eq!(
-            state.checkpoints(),
-            &BTreeMap::from([(1_u64, 12_u64), (2_u64, 20_u64), (3_u64, 35_u64)])
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_decision_labels_are_stable() {
-        let advance = FlowCheckpointDecision::AdvancedIncremental {
-            participating_regions: 1,
-            watermarks: 1,
-        };
-        let fallback = FlowCheckpointDecision::FallbackToFullSnapshot {
-            previous_mode: CheckpointMode::Incremental,
-            reason: FlowQueryFallbackReason::StaleCursor,
-        };
-
-        assert_eq!(advance.mode_label(), "incremental");
-        assert_eq!(advance.decision_label(), CHECKPOINT_DECISION_ADVANCE);
-        assert_eq!(advance.reason_label(), CHECKPOINT_REASON_NONE);
-        assert_eq!(fallback.mode_label(), "incremental");
-        assert_eq!(fallback.decision_label(), CHECKPOINT_DECISION_FALLBACK);
-        assert_eq!(fallback.reason_label(), "stale_cursor");
-    }
-
-    #[tokio::test]
-    async fn test_incremental_failure_falls_back_to_full_snapshot() {
-        let task = new_test_task_with_missing_sink().await;
-        task.state
-            .write()
-            .unwrap()
-            .advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-
-        let is_stale_cursor = task.handle_flow_query_failure(&ordinary_query_error(), None);
-
-        assert!(!is_stale_cursor);
-        let state = task.state.read().unwrap();
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert_eq!(state.checkpoints(), &BTreeMap::from([(1_u64, 10_u64)]));
-    }
-
-    #[tokio::test]
-    async fn test_incremental_failure_without_query_scope_marks_full_flow_dirty() {
-        let task = new_test_task_with_missing_sink().await;
-        {
-            let mut state = task.state.write().unwrap();
-            state.advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-            state.dirty_time_windows.clean();
-        }
-
-        let is_stale_cursor = task.handle_flow_query_failure(&ordinary_query_error(), None);
-
-        assert!(!is_stale_cursor);
-        let state = task.state.read().unwrap();
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert_eq!(state.dirty_time_windows.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_executed_query_failure_restores_scoped_dirty_windows_for_flush_path() {
-        let (task, plan) = new_test_task_and_plan_with_missing_sink().await;
-        {
-            let mut state = task.state.write().unwrap();
-            state.advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-            state.dirty_time_windows.clean();
-        }
-        let scoped_query = PlanInfo {
-            plan,
-            filter: Some(FilterExprInfo {
-                expr: datafusion_expr::lit(true),
-                col_name: "ts".to_string(),
-                time_ranges: vec![(Timestamp::new_second(10), Timestamp::new_second(20))],
-                window_size: chrono::Duration::seconds(10),
-            }),
-        };
-
-        let is_stale_cursor =
-            task.handle_executed_query_failure(&ordinary_query_error(), Some(&scoped_query));
-
-        assert!(!is_stale_cursor);
-        let state = task.state.read().unwrap();
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert_eq!(state.dirty_time_windows.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_stale_cursor_failure_restores_scoped_dirty_windows_only() {
-        let (task, plan) = new_test_task_and_plan_with_missing_sink().await;
-        {
-            let mut state = task.state.write().unwrap();
-            state.advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-            state.dirty_time_windows.clean();
-        }
-        let scoped_query = PlanInfo {
-            plan,
-            filter: Some(FilterExprInfo {
-                expr: datafusion_expr::lit(true),
-                col_name: "ts".to_string(),
-                time_ranges: vec![(Timestamp::new_second(10), Timestamp::new_second(20))],
-                window_size: chrono::Duration::seconds(10),
-            }),
-        };
-
-        let is_stale_cursor =
-            task.handle_executed_query_failure(&stale_cursor_query_error(), Some(&scoped_query));
-
-        assert!(is_stale_cursor);
-        let state = task.state.read().unwrap();
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-        assert_eq!(state.checkpoints(), &BTreeMap::from([(1_u64, 10_u64)]));
-        assert_eq!(state.dirty_time_windows.len(), 1);
-        assert_eq!(
-            state.dirty_time_windows.window_size(),
-            std::time::Duration::from_secs(10)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_build_flow_query_extensions_switches_with_checkpoint_mode() {
-        let task = new_test_task_with_missing_sink().await;
-
-        let extensions = task.build_flow_query_extensions(false).await.unwrap();
-        assert_eq!(
-            extensions,
-            vec![("flow.return_region_seq", "true".to_string())]
-        );
-
-        task.state
-            .write()
-            .unwrap()
-            .advance_checkpoints(HashMap::from([(1_u64, 10_u64), (2_u64, 20_u64)]));
-
-        let extensions = task.build_flow_query_extensions(false).await.unwrap();
-        assert!(extensions.contains(&("flow.return_region_seq", "true".to_string())));
-        assert!(
-            !extensions
-                .iter()
-                .any(|(key, _)| *key == FLOW_INCREMENTAL_MODE)
-        );
-        assert!(
-            !extensions
-                .iter()
-                .any(|(key, _)| *key == FLOW_INCREMENTAL_AFTER_SEQS)
-        );
-
-        let extensions = task.build_flow_query_extensions(true).await.unwrap();
-
-        assert!(extensions.contains(&("flow.return_region_seq", "true".to_string())));
-        assert!(extensions.contains(&(
-            FLOW_INCREMENTAL_MODE,
-            FLOW_INCREMENTAL_MODE_MEMTABLE_ONLY.to_string()
-        )));
-        assert!(extensions.contains(&(
-            FLOW_INCREMENTAL_AFTER_SEQS,
-            serde_json::json!({"1": 10, "2": 20}).to_string(),
-        )));
-
-        task.state.write().unwrap().disable_incremental();
-        let extensions = task.build_flow_query_extensions(true).await.unwrap();
-        assert!(extensions.contains(&("flow.return_region_seq", "true".to_string())));
-        assert!(
-            !extensions
-                .iter()
-                .any(|(key, _)| *key == FLOW_INCREMENTAL_MODE)
-        );
-        assert!(
-            !extensions
-                .iter()
-                .any(|(key, _)| *key == FLOW_INCREMENTAL_AFTER_SEQS)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_prepare_plan_for_incremental_disables_on_non_aggregate() {
-        let query_engine = create_test_query_engine();
-        let ctx = QueryContext::arc();
-        let plan = sql_to_df_plan(
-            ctx.clone(),
-            query_engine.clone(),
-            "SELECT number, ts FROM numbers_with_ts",
-            true,
-        )
-        .await
-        .unwrap();
-
-        // Build a DML wrapper using a real sink table from the test engine.
-        let (sink_table, _) = get_table_info_df_schema(
-            query_engine.engine_state().catalog_manager().clone(),
-            [
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ],
-        )
-        .await
-        .unwrap();
-        let table_provider = Arc::new(DfTableProviderAdapter::new(sink_table));
-        let table_source = Arc::new(DefaultTableSource::new(table_provider));
-        let dml_plan = LogicalPlan::Dml(DmlStatement::new(
-            datafusion_common::TableReference::bare("test"),
-            table_source,
-            WriteOp::Insert(datafusion_expr::dml::InsertOp::Append),
-            Arc::new(plan),
-        ));
-
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let task = BatchingTask::try_new(TaskArgs {
-            flow_id: 1,
-            query: "SELECT number, ts FROM numbers_with_ts",
-            plan: dml_plan.clone(),
-            time_window_expr: None,
-            expire_after: None,
-            sink_table_name: [
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ],
-            source_table_names: vec![[
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ]],
-            query_ctx: ctx,
-            catalog_manager: query_engine.engine_state().catalog_manager().clone(),
-            shutdown_rx: rx,
-            batch_opts: Arc::new(BatchingModeOptions::default()),
-            flow_eval_interval: None,
-        })
-        .unwrap();
-
-        // Put the state into Incremental mode with checkpoints.
-        task.state
-            .write()
-            .unwrap()
-            .advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-        assert_eq!(
-            task.state.read().unwrap().checkpoint_mode(),
-            CheckpointMode::Incremental
-        );
-
-        let (_plan, incremental_safe) = task
-            .prepare_plan_for_incremental(&dml_plan, None)
-            .await
-            .unwrap();
-        assert!(!incremental_safe);
-        let state = task.state.read().unwrap();
-        assert!(state.is_incremental_disabled());
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-    }
-
-    #[tokio::test]
-    async fn test_prepare_plan_for_incremental_disables_on_rewrite_error() {
-        let query_engine = create_test_query_engine();
-        let ctx = QueryContext::arc();
-        let plan = sql_to_df_plan(
-            ctx.clone(),
-            query_engine.clone(),
-            "SELECT sum(number) AS total, ts FROM numbers_with_ts GROUP BY ts",
-            true,
-        )
-        .await
-        .unwrap();
-
-        let (sink_table, _) = get_table_info_df_schema(
-            query_engine.engine_state().catalog_manager().clone(),
-            [
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ],
-        )
-        .await
-        .unwrap();
-        let table_provider = Arc::new(DfTableProviderAdapter::new(sink_table));
-        let table_source = Arc::new(DefaultTableSource::new(table_provider));
-        let dml_plan = LogicalPlan::Dml(DmlStatement::new(
-            datafusion_common::TableReference::bare("test"),
-            table_source,
-            WriteOp::Insert(datafusion_expr::dml::InsertOp::Append),
-            Arc::new(plan),
-        ));
-
-        let (_tx, rx) = tokio::sync::oneshot::channel();
-        let task = BatchingTask::try_new(TaskArgs {
-            flow_id: 1,
-            query: "SELECT sum(number) AS total, ts FROM numbers_with_ts GROUP BY ts",
-            plan: dml_plan.clone(),
-            time_window_expr: None,
-            expire_after: None,
-            // The sink table exists, but does not have the rewritten aggregate
-            // output column `total`, so the rewrite fails deterministically.
-            sink_table_name: [
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ],
-            source_table_names: vec![[
-                "greptime".to_string(),
-                "public".to_string(),
-                "numbers_with_ts".to_string(),
-            ]],
-            query_ctx: ctx,
-            catalog_manager: query_engine.engine_state().catalog_manager().clone(),
-            shutdown_rx: rx,
-            batch_opts: Arc::new(BatchingModeOptions::default()),
-            flow_eval_interval: None,
-        })
-        .unwrap();
-
-        task.state
-            .write()
-            .unwrap()
-            .advance_checkpoints(HashMap::from([(1_u64, 10_u64)]));
-        assert_eq!(
-            task.state.read().unwrap().checkpoint_mode(),
-            CheckpointMode::Incremental
-        );
-
-        let (_plan, incremental_safe) = task
-            .prepare_plan_for_incremental(&dml_plan, None)
-            .await
-            .unwrap();
-        assert!(!incremental_safe);
-        let state = task.state.read().unwrap();
-        assert!(state.is_incremental_disabled());
-        assert_eq!(state.checkpoint_mode(), CheckpointMode::FullSnapshot);
-    }
-
-    #[tokio::test]
-    async fn test_gen_create_table_sql() {
-        let query_engine = create_test_query_engine();
-        let ctx = QueryContext::arc();
-        struct TestCase {
-            sql: String,
-            sink_table_name: String,
-            column_schemas: Vec<ColumnSchema>,
-            primary_keys: Vec<String>,
-            time_index: String,
-        }
-
-        let update_at_schema = ColumnSchema::new(
-            AUTO_CREATED_UPDATE_AT_TS_COL,
-            ConcreteDataType::timestamp_millisecond_datatype(),
-            true,
-        );
-
-        let ts_placeholder_schema = ColumnSchema::new(
-            AUTO_CREATED_PLACEHOLDER_TS_COL,
-            ConcreteDataType::timestamp_millisecond_datatype(),
-            false,
-        )
-        .with_time_index(true);
-
-        let testcases = vec![
-            TestCase {
-                sql: "SELECT number, ts FROM numbers_with_ts".to_string(),
-                sink_table_name: "new_table".to_string(),
-                column_schemas: vec![
-                    ColumnSchema::new("number", ConcreteDataType::uint32_datatype(), true),
-                    ColumnSchema::new(
-                        "ts",
-                        ConcreteDataType::timestamp_millisecond_datatype(),
-                        false,
-                    )
-                    .with_time_index(true),
-                    update_at_schema.clone(),
-                ],
-                primary_keys: vec![],
-                time_index: "ts".to_string(),
-            },
-            TestCase {
-                sql: "SELECT number, max(ts) FROM numbers_with_ts GROUP BY number".to_string(),
-                sink_table_name: "new_table".to_string(),
-                column_schemas: vec![
-                    ColumnSchema::new("number", ConcreteDataType::uint32_datatype(), true),
-                    ColumnSchema::new(
-                        "max(numbers_with_ts.ts)",
-                        ConcreteDataType::timestamp_millisecond_datatype(),
-                        true,
-                    ),
-                    update_at_schema.clone(),
-                    ts_placeholder_schema.clone(),
-                ],
-                primary_keys: vec!["number".to_string()],
-                time_index: AUTO_CREATED_PLACEHOLDER_TS_COL.to_string(),
-            },
-            TestCase {
-                sql: "SELECT max(number), ts FROM numbers_with_ts GROUP BY ts".to_string(),
-                sink_table_name: "new_table".to_string(),
-                column_schemas: vec![
-                    ColumnSchema::new(
-                        "max(numbers_with_ts.number)",
-                        ConcreteDataType::uint32_datatype(),
-                        true,
-                    ),
-                    ColumnSchema::new(
-                        "ts",
-                        ConcreteDataType::timestamp_millisecond_datatype(),
-                        false,
-                    )
-                    .with_time_index(true),
-                    update_at_schema.clone(),
-                ],
-                primary_keys: vec![],
-                time_index: "ts".to_string(),
-            },
-            TestCase {
-                sql: "SELECT number, ts FROM numbers_with_ts GROUP BY ts, number".to_string(),
-                sink_table_name: "new_table".to_string(),
-                column_schemas: vec![
-                    ColumnSchema::new("number", ConcreteDataType::uint32_datatype(), true),
-                    ColumnSchema::new(
-                        "ts",
-                        ConcreteDataType::timestamp_millisecond_datatype(),
-                        false,
-                    )
-                    .with_time_index(true),
-                    update_at_schema.clone(),
-                ],
-                primary_keys: vec!["number".to_string()],
-                time_index: "ts".to_string(),
-            },
-        ];
-
-        for tc in testcases {
-            let plan = sql_to_df_plan(ctx.clone(), query_engine.clone(), &tc.sql, true)
-                .await
-                .unwrap();
-            let expr = create_table_with_expr(
-                &plan,
-                &[
-                    "greptime".to_string(),
-                    "public".to_string(),
-                    tc.sink_table_name.clone(),
-                ],
-                &QueryType::Sql,
-            )
-            .unwrap();
-            // TODO(discord9): assert expr
-            let column_schemas = expr
-                .column_defs
-                .iter()
-                .map(|c| try_as_column_schema(c).unwrap())
-                .collect::<Vec<_>>();
-            assert_eq!(tc.column_schemas, column_schemas, "{:?}", tc.sql);
-            assert_eq!(tc.primary_keys, expr.primary_keys, "{:?}", tc.sql);
-            assert_eq!(tc.time_index, expr.time_index, "{:?}", tc.sql);
-        }
-    }
-}
+mod test;
